test(error): Extract error assertion
